fix: Added Global exception handling for Book

+++ src/main/java/pk/edu/niit/library_management_system/Book/Services/BookServices.java
@@ -6,6 +6,7 @@ import org.springframework.stereotype.Service;
 import pk.edu.niit.library_management_system.Author.Repository.AuthorRepository;
 import pk.edu.niit.library_management_system.Book.Entity.Book;
 import pk.edu.niit.library_management_system.Book.Repository.BookRepository;
+import pk.edu.niit.library_management_system.ExceptionHandler.BookNotFoundException;
 
 import java.util.List;
 import java.util.Optional;
@@ -18,6 +19,7 @@ public class BookServices {
 
     public List<Book> getAllBooks()
     {
+
         return bookRepository.findAll();
     }
 
@@ -28,34 +30,32 @@ public class BookServices {
 
     public void deleteBook(long id)
     {
+        if(!bookRepository.existsById(id))
+        {
+            throw new BookNotFoundException("DELETE/book/id/"+id+":Book not found for this id "+id);
+        }
         bookRepository.deleteById(id);
     }
 
     public Book updateBook(long id,Book book)
     {
-        Optional<Book> existing=bookRepository.findById(id);
-        if(existing.isPresent())
-        {
-            Book updated=existing.get();
-            updated.setIsbn(book.getIsbn());
-            updated.setTitle(book.getTitle());
-            updated.setAuthor(book.getAuthor());
-            updated.setTotalCopies(book.getTotalCopies());
-            updated.setAvailableCopies(book.getAvailableCopies());
-            return bookRepository.save(updated);
-        }
-        return null;
+       Book existing=bookRepository.findById(id).orElseThrow(()-> new BookNotFoundException(
+               "PUT/book/id/"+id+": Book not found for id : "+id));
+
+            existing.setIsbn(book.getIsbn());
+            existing.setTitle(book.getTitle());
+            existing.setAuthor(book.getAuthor());
+            existing.setTotalCopies(book.getTotalCopies());
+            existing.setAvailableCopies(book.getAvailableCopies());
+            return bookRepository.save(existing);
+
     }
 
     public Book getBookByID(long id)
     {
-        Optional<Book> book=bookRepository.findById(id);
-        if (book.isPresent())
-        {
-            Book bookFound=book.get();
-            return bookFound;
-        }
-        return null;
+        Book book=bookRepository.findById(id).orElseThrow(()->new BookNotFoundException("PUT/book/id/"+id+": " +
+                "Book not found for id : "+id));
+        return book;
 
     }
 }
